Updated SQL data insert and started adding basic CRUD for Accounts

+++ src/main/java/com/onlydan/od/entities/OrderDetails.java
@@ -25,7 +25,7 @@ public class OrderDetails {
     private ProductInfo productId;
 
     @Column(precision = 10, scale = 2)
-    private Integer productPrice;
+    private Double productPrice;
 
     @Column
     private Boolean isOnSale;
